% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thoughts_prayers.R
\name{thoughts_prayers}
\alias{thoughts_prayers}
\title{Thoughts and Prayers}
\usage{
thoughts_prayers(...)
}
\arguments{
\item{...}{whatever you want. It doesn't matter}
}
\description{
This function takes anything you want and returns NULL, just like actual thoughts and prayers. Do better
}
\examples{
thoughts_prayers("whatever, man")
}
